flink-app/q1-job/requirements.txt
apache-flink==1.17.1
requests
pillow
numpy
umsgpack

